feat: enhance crearProveedor and crearProveedorArticulo to set default supplier-article relationships and improve response messages

+++ src/validators/articulo.schema.js
@@ -2,11 +2,23 @@ import { z } from 'zod'
 
 export const articuloSchema = z.object({
   descripcion: z.string().min(1, { message: 'La descripción es obligatoria' }),
-  demanda_articulo: z.number({ invalid_type_error: 'Debe ser un número' }).int().nonnegative({ message: 'No puede ser negativo' }),
-  desviacion_est_dem: z.number({ invalid_type_error: 'Debe ser un número' }).int().nonnegative({ message: 'No puede ser negativo' }),
-  costo_almacenamiento: z.number({ invalid_type_error: 'Debe ser un número' }).nonnegative({ message: 'No puede ser negativo' }),
-  stock: z.number({ invalid_type_error: 'Debe ser un número' }).int().nonnegative({ message: 'No puede ser negativo' }),
-  precioVenta: z.number({ invalid_type_error: 'Debe ser un número' }).nonnegative({ message: 'No puede ser negativo' }),
+  demanda_articulo: z
+    .number({ invalid_type_error: 'Debe ser un número' })
+    .int()
+    .nonnegative({ message: 'No puede ser negativo' }),
+  desviacion_est_dem: z
+    .number({ invalid_type_error: 'Debe ser un número' })
+    .nonnegative({ message: 'No puede ser negativo' }),
+  costo_almacenamiento: z
+    .number({ invalid_type_error: 'Debe ser un número' })
+    .nonnegative({ message: 'No puede ser negativo' }),
+  stock: z
+    .number({ invalid_type_error: 'Debe ser un número' })
+    .int()
+    .nonnegative({ message: 'No puede ser negativo' }),
+  precioVenta: z
+    .number({ invalid_type_error: 'Debe ser un número' })
+    .nonnegative({ message: 'No puede ser negativo' }),
 })
 
 export function validateArticulo(object) {
